Fix jerky scrolling on sending
Maybe you never noticed but `discardButton` is always visible at
the beginning and there is a chance that the height of
`Item#message` will be based on its height. But when it gets
invisible shortly after in normal circumstances it could cause
jerky scrolling.

This has been introduced by 87b18d9e via changing `ActiveLabel` to
`ToolButton` which is taller. This patch applies a workaround which
prevents showing it immediately in full height.